[libc] Cleaup which entrypoints are fullbuild only
This patch moves some functions in/out of fullbuild based on if it's
necessary. Specifically:
dlfcn.h: dynamic linker specific, currently stubs, now fullbuild only
arpa/inet.h: The currently implemented functions are standalone
sys/select.h and socket.h: Syscall wrappers, safe to overlay

There's also some cleanup to make sure everything builds.

+++ libc/test/src/sys/select/select_ui_test.cpp
@@ -11,7 +11,7 @@
 #include "src/unistd/read.h"
 #include "test/UnitTest/Test.h"
 
-#include <sys/select.h>
+#include <sys/select.h> //TODO: move to hdr/sys_select_macros.h
 #include <unistd.h>
 
 // This test is not be run automatically as part of the libc testsuite.
@@ -23,26 +23,22 @@ TEST(LlvmLibcSelectTest, ReadStdinAfterSelect) {
   fd_set set;
   FD_ZERO(&set);
   FD_SET(STDIN_FD, &set);
-  struct timeval zero {
-    0, 0
-  }; // No wait
-  struct timeval hr {
-    3600, 0
-  }; // Wait for an hour.
+  struct timeval zero{0, 0};  // No wait
+  struct timeval hr{3600, 0}; // Wait for an hour.
 
   // Zero timeout means we don't wait for input. So, select should return
   // immediately.
   int count =
       LIBC_NAMESPACE::select(STDIN_FD + 1, &set, nullptr, nullptr, &zero);
   // The set should indicate that stdin is NOT ready for reading.
-  ASSERT_EQ(0, FD_ISSET(STDIN_FD, &set));
+  ASSERT_EQ(0, static_cast<int>(FD_ISSET(STDIN_FD, &set)));
 
   FD_SET(STDIN_FD, &set);
   // Wait for an hour and give the user a chance to hit a key.
   count = LIBC_NAMESPACE::select(STDIN_FD + 1, &set, nullptr, nullptr, &hr);
   ASSERT_EQ(count, 1);
   // The set should indicate that stdin is ready for reading.
-  ASSERT_EQ(1, FD_ISSET(STDIN_FD, &set));
+  ASSERT_EQ(1, static_cast<int>(FD_ISSET(STDIN_FD, &set)));
 
   // Verify that atleast one character can be read.
   char c;
